requirements.txt
pycrypto==2.6.1
aiohttp==2.0.7
async-timeout==1.2.0
chardet==3.0.2
multidict==2.1.4
yarl==0.10.0
websockets==3.3
aiohttp_jinja2==0.13.0
Jinja2==2.9.6
MarkupSafe==1.0
